% Generated by roxygen2 (4.0.2): do not edit by hand
\name{display.xterm.colors}
\alias{display.ansi.colors}
\alias{display.xterm.colors}
\title{Display color tables}
\usage{
display.xterm.colors(perm = 1:3)

display.ansi.colors(numbers = TRUE)
}
\arguments{
\item{perm}{Rotation of the colour cube, supplied as a permutation of its
dimensions. Sent to \code{\link[base]{aperm}}.}

\item{numbers}{Logical, whether to display colour indices.}
}
\value{
Nothing
}
\description{
The xterm colour table consist of the ANSI colours (16), the web
colour cube (216) and additional shades of grey not including full white and
black (16). However these are not strictly defined but can vary somewhat
between systems and configurations.
}
\examples{
display.xterm.colors()
display.ansi.colors()
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
\seealso{
style
}

