\name{display.xterm.pal}
\alias{display.xterm.pal}
\title{Get predefined colour palettes}
\usage{
  display.xterm.pal(pal, numbers = FALSE)
}
\arguments{
  \item{pal}{Palette name(s). Leave blank for all.}

  \item{numbers}{Whether to show colour indices.}
}
\value{
  Nothing
}
\description{
  Get predefined colour palettes
}
\examples{
display.xterm.pal("BuPu")
}
\author{
  Christofer \enc{Bäcklin}{Backlin}
}

