\name{style.auto.demo}
\alias{style.auto.demo}
\title{Demonstration of automatic styles.}
\usage{
  style.auto.demo()
}
\value{
  Nothing
}
\description{
  Demonstration of automatic styles.
}
\examples{
style.auto.demo()
}
\author{
  Christofer \enc{Bäcklin}{Backlin}
}
\seealso{
  style.auto, style
}

