\name{style.light}
\alias{style.dark}
\alias{style.light}
\title{Change color theme}
\usage{
  style.light()

  style.dark()
}
\value{
  Nothing
}
\description{
  Light is designed for light backgrounds and dark for dark
  backgrounds. This function merely sets a global option
  (\code{color.scheme}).
}
\examples{
style.light()
style.dark()
}
\author{
  Christofer \enc{Bäcklin}{Backlin}
}
\seealso{
  style.auto
}

