% Generated by roxygen2 (4.0.2): do not edit by hand
\name{style.palette}
\alias{style.palette}
\title{Define the palette for auto-styling}
\usage{
style.palette(x = c("dark", "light"))
}
\arguments{
\item{x}{Palette name or definition.}
}
\value{
Nothing, sets the approprite \code{\link{options}} variables.
}
\description{
Define the palette for auto-styling
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
\seealso{
style.auto
}

