% Generated by roxygen2 (4.0.2): do not edit by hand
\name{style}
\alias{style}
\alias{style.clear}
\alias{style.off}
\title{Color termnal output}
\usage{
style(..., fg = NA, bg = NA, font = NA, mode = c("xterm-256color",
  "ansi", "off"))

style.clear(...)

style.off()
}
\arguments{
\item{...}{Sent to \code{\link{cat}}.}

\item{fg}{Foreground color i.e. color of the text. Can be any number in
[0, 255] or a string such as \code{"grey"} or \code{"dark red"} (for the
basic 16 colors).}

\item{bg}{Background color. Takes the same values as \code{fg}.}

\item{font}{A vector containg any combination of the following font styles:
\code{"normal"}, \code{"bold"}, \code{"underline"},
\code{"blink"} (renders as bold on most terminals), \code{"inverse"}.
Note that these may not be rendered on all terminals.}

\item{mode}{Escape code mode.}
}
\value{
Nothing, sends all output to \code{\link{cat}}.
}
\description{
Talk about the different options. style.mode, style.palette.
}
\examples{
style("Blue suede shoes\\n", bg="blue")
style(fg="red")
cat("everything is red now!")
style(NULL)
cat("but not anymore!")
}
\author{
Christofer \enc{Bäcklin}{Backlin}

Christofer \enc{Bäcklin}{Backlin}
}
\seealso{
style
}

