\name{style.set}
\alias{style.set}
\title{Set, get and clear style.}
\usage{style.set(fg, bg, font.style, make.default=TRUE)
style.clear(make.default=TRUE)
}
\description{Set, get and clear style.}
\details{\code{style.set}: Generate formatting escape sequences from parameters.

}
\value{\code{style.set}: Formatting escape sequences.

\code{style.get}: The formatting escape sequence generated by the last call to
\code{style.set} (except those called with \code{make.default = FALSE},
e.g. all package internal calls).


\code{style.clear}: Formatting escape sequence for reseting style to default,
\code{\\033[0m}.


}
\author{Christofer Bäcklin}
\seealso{\code{\link{style}}}
\arguments{\item{fg}{Foreground color i.e. color of the text. Can be any number in
[0, 255] or a string such as \code{"grey"} or \code{"dark red"} (for the
basic 16 colors).}
\item{bg}{Background color. Takes the same values as \code{fg}.}
\item{font.style}{A vector containg any combination of the following
elements: \code{"normal"}, \code{"bold"}, \code{"underline"},
\code{"blink"} (renders as bold on most terminals), \code{"inverse"}.}
\item{make.default}{If \code{TRUE} (default) this style will be the one
reverted to after temporary changes in style e.g. with the \code{style}
command or any display function like \code{\link[dataview]{entry.view}}.}
}
\alias{style.get}
\alias{style.clear}

