% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/colors.R
\name{xterm256.to.ansi}
\alias{xterm256.to.ansi}
\title{Convert xterm256 color code to ansi code}
\usage{
xterm256.to.ansi(x)
}
\arguments{
\item{x}{Integer specifying xterm256 color.}
}
\value{
Integer that approximates \code{x} in the ansi palette.
}
\description{
Convert xterm256 color code to ansi code
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}

