% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.xtermStyle}
\alias{print.xtermStyle}
\title{Print a style object}
\usage{
\method{print}{xtermStyle}(x, ...)
}
\arguments{
\item{x}{\code{\link{style}} object.}

\item{...}{Ignored, kept for S3 consistency.}
}
\description{
Print a style object
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
\seealso{
style
}

