% Generated by roxygen2 (4.0.2): do not edit by hand
\name{style.auto}
\alias{style.auto}
\title{Automatic styling according to object properties.}
\usage{
style.auto(x, ...)
}
\arguments{
\item{x}{Object to decide formatting from.}

\item{\dots}{Sent to \code{\link{style}}.}
}
\value{
See \code{\link{style}}.
}
\description{
Automatic styling according to object properties.
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
\seealso{
style, style.palette
}

