% Generated by roxygen2 (4.0.2): do not edit by hand
\name{style.palette}
\alias{get.palette}
\alias{style.palette}
\title{Define the palette for auto-styling}
\usage{
style.palette(x, ...)

get.palette(x = c("dark", "light"))
}
\arguments{
\item{x}{Palette name or definition.}

\item{...}{Settings that override the defaults.}
}
\value{
Nothing, sets the approprite \code{\link{options}} variables.
}
\description{
Define the palette for auto-styling
}
\examples{
style.palette("light")
style.palette(list(fg=c(numeric=10, character=13)))
style.palette("light", fg=c(numeric=10, character=13))
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
\seealso{
style.auto, display.xterm.colors
}

