\name{style.set}
\alias{style.clear}
\alias{style.get}
\alias{style.set}
\title{Set, get and clear style.}
\usage{
  style.set(fg = NULL, bg = NULL, font.style = NULL,
    make.default = TRUE)

  style.get()

  style.clear(make.default = TRUE)
}
\arguments{
  \item{fg}{Foreground color i.e. color of the text. Can be
  any number in [0, 255] or a string such as \code{"grey"}
  or \code{"dark red"} (for the basic 16 colors).}

  \item{bg}{Background color. Takes the same values as
  \code{fg}.}

  \item{font.style}{A vector containg any combination of
  the following elements: \code{"normal"}, \code{"bold"},
  \code{"underline"}, \code{"blink"} (renders as bold on
  most terminals), \code{"inverse"}.}

  \item{make.default}{If \code{TRUE} (default) this style
  will be the one reverted to after temporary changes in
  style e.g. with the \code{style} command or any display
  function like \code{\link[dataview]{entry.view}}.}
}
\value{
  Formatting escape sequences.

  The formatting escape sequence generated by the last call
  to \code{style.set} (except those called with
  \code{make.default = FALSE}, e.g. all package internal
  calls).

  Formatting escape sequence for reseting style to default,
  \code{\\033[0m}.
}
\description{
  Generate formatting escape sequences from parameters.

  Get current default style.

  Clear current style to terminal default.
}
\author{
  Christofer \enc{Bäcklin}{Backlin}

  Christofer \enc{Bäcklin}{Backlin}

  Christofer \enc{Bäcklin}{Backlin}
}
\seealso{
  style
}

