% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/colors.r
\name{xterm256.to.ANSI}
\alias{xterm256.to.ANSI}
\title{Convert xterm256 color code to ANSI code}
\usage{
xterm256.to.ANSI(x)
}
\arguments{
\item{x}{Integer specifying xterm256 color.}
}
\value{
Integer that approximates \code{x} in the ANSI palette.
}
\description{
Convert xterm256 color code to ANSI code
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}

